requirements.txt
Django==6.0.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
asgiref==3.11.1
psycopg==3.3.3
psycopg-binary==3.3.3
joblib==1.5.3
numpy
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
